% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{Convert Integer to Fizz Buzz.}
\usage{
FizzBuzz(x)
}
\arguments{
\item{\code{x}}{A vector of integers}
}
\value{
The Fizz Buzz-ification of \code{x}
}
\description{
When entering a vector of integers of any length into the function, the
output will show which integers are divisible by which numbers.
}
\details{
If an integer is divisible by three, it will be converted to "Fizz".
If an integer is divisible by five, it will be converted to "Buzz".
If an integer is divisible by both three and five, then it will be converted
to "Fizz Buzz".
}
\examples{
smallVector <- c(1:30)
largerVector <- sample(x = c(1:10), size = 90, replace = TRUE)

FizzBuzz(smallVector)
FizzBuzz(largerVector)
}
