% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PulliamAirportWeather.R
\docType{data}
\name{PulliamAirportWeather}
\alias{PulliamAirportWeather}
\title{A time series of various weather variables at the Flagstaff Pulliam Airport.}
\format{
a data frame with 22967 observations
\describe{
  \item{date}{The date of observation as a POSIX date format.}
  \item{prcp}{Daily percipitation in inches.}
  \item{snow}{Daily snow fall in inches}
  \item{tmax}{Daily maximum temperature measured in Fahrenheit}
  \item{tmin}{Daily minimum temperature measured in Fahrenheit}
}
}
\source{
\url{www.ncdc.noaa.gov}
}
\usage{
PulliamAirportWeather
}
\description{
A time series of various weather variables at the Flagstaff Pulliam Airport.
}
\keyword{datasets}
